Calculate Gamma function

It calculates the value of the integral defined in Definition 11 from references. It is implementation of the Theorem 8 from references and is using the formula (19) from references.

Usage

calculate_gamma_function(perm, lambda)

Arguments

perm

An object of a gips_perm class.

lambda

A positive real number.

Value

Returns the value of the Gamma function.

References

Piotr Graczyk, Hideyuki Ishi, Bartosz Kolodziejek, Hélène Massam. "Model selection in the space of Gaussian models invariant by symmetry." The Annals of Statistics, 50(3) 1747-1774 June 2022. DOI: 10.1214/22-AOS2174

See also

Examples

id_perm <- gips_perm(permutations::id, 2)
calculate_gamma_function(id_perm, 0.5001) # 10.7...
#> [1] 10.70139
calculate_gamma_function(id_perm, 0.50000001) # 19.9...
#> [1] 19.91198
calculate_gamma_function(id_perm, 0.500000000001) # 29.1...
#> [1] 29.12235
# calculate_gamma_function(id_perm, 0.5) # integral diverges; returns Inf and warning