Prepare orthogonal matrix

Calculate orthogonal matrix U_Gamma for decomposition in Theorem 1 from references.

Usage

prepare_orthogonal_matrix(perm, perm_size = NULL, basis = NULL)

Arguments

perm

Object of a gips_perm or a permutations::cycle class.

perm_size

Size of permutation. Required if perm is of a permutations::cycle class.

basis

A matrix with basis vectors in COLUMNS. Identity by default.

Value

A matrix perm_size x perm_size with columns from V object elements, sorted according to Theorem 6.

Details

Given X - invariant under the permutation perm. Call Gamma the permutations cyclic group \(<perm> = {perm, perm^2, ...}\).

Then, U_Gamma is such an orthogonal matrix that X is "pretty" in it.

To be more precise, the matrix t(U_Gamma) %*% X %*% U_Gamma has a lot of zeros (see examples).

To calculate it, Theorem 6 from references is used.

References

Piotr Graczyk, Hideyuki Ishi, Bartosz Kolodziejek, Hélène Massam. "Model selection in the space of Gaussian models invariant by symmetry." The Annals of Statistics, 50(3) 1747-1774 June 2022. DOI: 10.1214/22-AOS2174

See also

Examples

gperm <- gips_perm("(1,2,3)(4,5)", 5)
U_g <- prepare_orthogonal_matrix(gperm)

number_of_observations <- 10
X <- matrix(rnorm(5 * number_of_observations), number_of_observations, 5)
S <- cov(X)
X <- project_matrix(S, perm = gperm) # this matrix in invariant under gperm

t(U_g) %*% X %*% U_g # the non-zeros only on diagonal and [1,2] and [2,1]
#>               [,1]          [,2]          [,3]         [,4]      [,5]
#> [1,]  5.810968e-01  1.231035e-01 -8.326673e-17 5.551115e-17 0.0000000
#> [2,]  1.231035e-01  3.402060e-01 -1.734723e-17 1.387779e-17 0.0000000
#> [3,] -2.775558e-17 -1.962616e-17  7.596555e-01 2.220446e-16 0.0000000
#> [4,]  0.000000e+00  9.813078e-18  3.053113e-16 7.596555e-01 0.0000000
#> [5,]  0.000000e+00  0.000000e+00  0.000000e+00 0.000000e+00 0.2126789