Summarizing the gips object

summary method for class "gips".

Usage

# S3 method for gips
summary(object, ...)

# S3 method for summary.gips
print(x, ...)

Arguments

object

An object of class "gips"; is usually a result of a find_MAP().

...

Further arguments passed to or from other methods.

x

An object of class "summary.gips" to be printed

Value

The function summary.gips computes and returns a list of summary statistics of the given gips object. Those are:

For unoptimized gips object:

optimized - FALSE

start_permutation - the permutation this gips represents

start_permutation_log_posteriori - the log of the a posteriori value the start permutation has

n0 - the minimal number of observations needed for a MAP Estimator of the covariance matrix to exist

S_matrix - the underlying matrix; this is used to calculate the posteriori value

number_of_observations - the number of observations that were observed for the S_matrix to be calculated; this is used to calculate the posteriori value

delta, D_matrix - the parameters of the Bayesian method

For optimized gips object:

optimized - TRUE

found_permutation - the permutation this gips represents; the visited permutation with the biggest a posteriori value

found_permutation_log_posteriori - the log of the a posteriori value the found permutation have

start_permutation - the original permutation this gips represented before optimization; the first visited permutation

start_permutation_log_posteriori - the log of the a posteriori value the start permutation has

n0 - the minimal number of observations needed for a MAP Estimator of the covariance matrix to exist

S_matrix - the underlying matrix; this is used to calculate the posteriori value

number_of_observations - the number of observations that were observed for the S_matrix to be calculated; this is used to calculate the posteriori value

delta, D_matrix - the parameters of the Bayesian method

optimization_algorithm_used - all used optimization algorithms in order (one could start optimization with "MH", and then do an "HC")

did_converge - boolean, did the last used algorithm converge

number_of_log_posteriori_calls - how many times was the log_posteriori_of_gips() function called during the optimization

full_optimization_time - how long was the optimization process; the sum of all optimization times (when there were multiple)

log_posteriori_calls_after_best - how many times was the log_posteriori_of_gips() function called after the found_permutation; in other words, how long ago could the optimization be stopped and have the same result; if this value is small, consider running find_MAP() one more time with optimizer = "continue"

acceptance_rate - only interesting for optimizer = "MH"; how often was the algorithm accepting the change of permutation in an iteration

Methods (by generic)

print(summary.gips): Printing method for class "summary.gips". Prints every interesting information in a pleasant, human readable form

See also

Examples

require("MASS") # for mvrnorm()

perm_size <- 6
mu <- numeric(perm_size)
sigma_matrix <- matrix(
  data = c(
    1.0, 0.8, 0.6, 0.4, 0.6, 0.8,
    0.8, 1.0, 0.8, 0.6, 0.4, 0.6,
    0.6, 0.8, 1.0, 0.8, 0.6, 0.4,
    0.4, 0.6, 0.8, 1.0, 0.8, 0.6,
    0.6, 0.4, 0.6, 0.8, 1.0, 0.8,
    0.8, 0.6, 0.4, 0.6, 0.8, 1.0
  ),
  nrow = perm_size, byrow = TRUE
) # sigma_matrix is a matrix invariant under permutation (1,2,3,4,5,6)
number_of_observations <- 13
Z <- MASS::mvrnorm(number_of_observations, mu = mu, Sigma = sigma_matrix)
S <- (t(Z) %*% Z) / number_of_observations # the theoretical mean is 0

g <- gips(S, number_of_observations)

g_map <- find_MAP(g, max_iter = 10, show_progress_bar = FALSE, optimizer = "MH")
unclass(summary(g_map))
#> $optimized
#> [1] TRUE
#> 
#> $found_permutation
#> [1] (36)
#> 
#> $found_permutation_log_posteriori
#> [1] -5.300003
#> 
#> $start_permutation
#> [1] ()
#> 
#> $start_permutation_log_posteriori
#> [1] -7.454616
#> 
#> $n0
#> [1] 5
#> 
#> $S_matrix
#>           [,1]      [,2]      [,3]      [,4]      [,5]      [,6]
#> [1,] 1.3551561 1.0006756 0.6889481 0.5309880 0.8854684 1.1971960
#> [2,] 1.0006756 0.8145939 0.5378129 0.4800485 0.6661302 0.9429112
#> [3,] 0.6889481 0.5378129 0.8330968 0.8260757 0.9772109 0.6819270
#> [4,] 0.5309880 0.4800485 0.8260757 1.1479401 1.1988796 0.8528523
#> [5,] 0.8854684 0.6661302 0.9772109 1.1988796 1.4182178 1.1071372
#> [6,] 1.1971960 0.9429112 0.6819270 0.8528523 1.1071372 1.3681214
#> 
#> $number_of_observations
#> [1] 13
#> 
#> $delta
#> [1] 3
#> 
#> $D_matrix
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]    1    0    0    0    0    0
#> [2,]    0    1    0    0    0    0
#> [3,]    0    0    1    0    0    0
#> [4,]    0    0    0    1    0    0
#> [5,]    0    0    0    0    1    0
#> [6,]    0    0    0    0    0    1
#> 
#> $optimization_algorithm_used
#> [1] "Metropolis_Hastings"
#> 
#> $did_converge
#> NULL
#> 
#> $number_of_log_posteriori_calls
#> [1] 10
#> 
#> $full_optimization_time
#> Time difference of 0.02220845 secs
#> 
#> $log_posteriori_calls_after_best
#> [1] 0
#> 
#> $acceptance_rate
#> [1] 0.1
#> 

g_map2 <- find_MAP(g, max_iter = 10, show_progress_bar = FALSE, optimizer = "HC")
summary(g_map2)
#> The optimized `gips` object.
#> 
#> Permutation:
#>  (3,6)
#> 
#> Log_posteriori:
#>  -5.300003
#> 
#> Number of observations:
#>  13
#> 
#> n0:
#>  5
#> 
#> Number of observations is bigger than n0 for this permutaion, so the gips model based on the found permutation does exist.
#> 
#> --------------------------------------------------------------------------------
#> Optimization algorithm:
#>  hill_climbing did converge
#> 
#> Number of log_posteriori calls:
#>  31
#> 
#> Optimization time:
#>  0.07141709 secs
#> 
#> Log_posteriori calls after the found permutation:
#>  18