Plot optimized matrix or optimization convergence

Plot method for gips object.

Usage

# S3 method for gips
plot(
  x,
  type = NA,
  logarithmic_y = TRUE,
  logarithmic_x = FALSE,
  color = NULL,
  title_text = "Convergence plot",
  xlabel = NULL,
  ylabel = NULL,
  show_legend = TRUE,
  ylim = NULL,
  xlim = NULL,
  ...
)

Arguments

x

Object of class gips. Has to first be optimized with find_MAP.

type

Character. A type of a plot. one of c("heatmap", "all", "best", "both"). For "heatmap", plots a heatmap of the S matrix inside the gips object that was projected on the permutation in the gips object. For "all", plots the line of the posterioris for all visited state. For "best", plots the line of the biggest posterioris up to the moment For "both", both lines from "all" and "best" are plotted. Default value is NA, which will be changed to "heatmap" for non-optimized gips objects, and to "both" for optimized ones. Using the default produces a warning. For the type = "heatmap", all other arguments are ignored.

logarithmic_y

boolean.

logarithmic_x

boolean.

color

Vector of colors to be used to plot lines.

title_text

Text to be in a title of the plot.

xlabel

Text to be on the bottom of the plot.

ylabel

Text to be on the left of the plot.

show_legend

boolean.

ylim

Limits of y axis. When NULL, the minimum and maximum of the log_posteriori_of_gips is taken.

xlim

Limits of x axis. When NULL, the whole optimization process is shown.

...

Additional arguments passed to heatmap or other various elements of the plot.

Value

Invisible NULL.