Summarizing the gips object

summary method for class "gips".

Usage

# S3 method for gips
summary(object, ...)

# S3 method for summary.gips
print(x, ...)

Arguments

object

An object of class "gips"; is usually a result of a find_MAP().

...

Further arguments passed to or from other methods.

x

An object of class "summary.gips" to be printed

Value

The function summary.gips computes and returns a list of summary statistics of the given gips object. Those are:

For unoptimized gips object:

optimized - FALSE

start_permutation - the permutation this gips represents

start_permutation_log_posteriori - the log of the a posteriori value the start permutation has

times_more_likely_than_id - how many more likely the start_permutation is over the identity permutation, (). It can be a number less than 1, which means the identity permutation, (), is more likely. Keep in mind this number can be really big and can be overflowed to Inf

n0 - the minimal number of observations needed for existence of the maximum likelihood estimator (corresponding to a MAP) of the covariance matrix (see \(C\sigma\) and n0 section in vignette("Theory", package = "gips") or in its pkgdown page).

S_matrix - the underlying matrix; this is used to calculate the posteriori value

number_of_observations - the number of observations that were observed for the S_matrix to be calculated; this is used to calculate the posteriori value

was_mean_estimated - given by the user while creating the gips object:

TRUE means the S parameter was output of stats::cov() function

FALSE means the S parameter was calculated with S = t(X) %*% X / number_of_observations

delta, D_matrix - the parameters of the Bayesian method

For optimized gips object:

optimized - TRUE

found_permutation - the permutation this gips represents; the visited permutation with the biggest a posteriori value

found_permutation_log_posteriori - the log of the a posteriori value the found permutation have

start_permutation - the original permutation this gips represented before optimization; the first visited permutation

start_permutation_log_posteriori - the log of the a posteriori value the start permutation has

times_more_likely_than_start - how many more likely the found_permutation is over the start_permutation. It cannot be a number less than 1. Keep in mind this number can be really big and can be overflowed to Inf

n0 - the minimal number of observations needed for existence of the maximum likelihood estimator (corresponding to a MAP) of the covariance matrix (see \(C\sigma\) and n0 section in vignette("Theory", package = "gips") or in its pkgdown page).

S_matrix - the underlying matrix; this is used to calculate the posteriori value

number_of_observations - the number of observations that were observed for the S_matrix to be calculated; this is used to calculate the posteriori value

was_mean_estimated - given by the user while creating the gips object:

TRUE means the S parameter was output of stats::cov() function

FALSE means the S parameter was calculated with S = t(X) %*% X / number_of_observations

delta, D_matrix - the parameters of the Bayesian method

optimization_algorithm_used - all used optimization algorithms in order (one could start optimization with "MH", and then do an "HC")

did_converge - a boolean, did the last used algorithm converge

number_of_log_posteriori_calls - how many times was the log_posteriori_of_gips() function called during the optimization

whole_optimization_time - how long was the optimization process; the sum of all optimization times (when there were multiple)

log_posteriori_calls_after_best - how many times was the log_posteriori_of_gips() function called after the found_permutation; in other words, how long ago could the optimization be stopped and have the same result; if this value is small, consider running find_MAP() one more time with optimizer = "continue". For optimizer = "BF", it is NULL

acceptance_rate - only interesting for optimizer = "MH"; how often was the algorithm accepting the change of permutation in an iteration

print.summary.gips returns an invisible NULL.

Methods (by generic)

print(summary.gips): Printing method for class "summary.gips". Prints every interesting information in a pleasant, human readable form

See also

Examples

require("MASS") # for mvrnorm()

perm_size <- 6
mu <- runif(6, -10, 10) # Assume we don't know the mean
sigma_matrix <- matrix(
  data = c(
    1.0, 0.8, 0.6, 0.4, 0.6, 0.8,
    0.8, 1.0, 0.8, 0.6, 0.4, 0.6,
    0.6, 0.8, 1.0, 0.8, 0.6, 0.4,
    0.4, 0.6, 0.8, 1.0, 0.8, 0.6,
    0.6, 0.4, 0.6, 0.8, 1.0, 0.8,
    0.8, 0.6, 0.4, 0.6, 0.8, 1.0
  ),
  nrow = perm_size, byrow = TRUE
) # sigma_matrix is a matrix invariant under permutation (1,2,3,4,5,6)
number_of_observations <- 13
Z <- MASS::mvrnorm(number_of_observations, mu = mu, Sigma = sigma_matrix)
S <- cov(Z) # Assume we have to estimate the mean

g <- gips(S, number_of_observations)

g_map <- find_MAP(g, max_iter = 10, show_progress_bar = FALSE, optimizer = "Metropolis_Hastings")
unclass(summary(g_map))
#> $optimized
#> [1] TRUE
#> 
#> $found_permutation
#> [1] ()
#> 
#> $found_permutation_log_posteriori
#> [1] -11.50795
#> 
#> $start_permutation
#> [1] ()
#> 
#> $start_permutation_log_posteriori
#> [1] -11.50795
#> 
#> $times_more_likely_than_start
#> [1] 1
#> 
#> $log_times_more_likely_than_start
#> [1] 0
#> 
#> $n0
#> [1] 7
#> 
#> $S_matrix
#>           [,1]      [,2]      [,3]      [,4]      [,5]      [,6]
#> [1,] 1.0596707 0.8553641 0.8715926 0.5600552 0.7643617 0.7481332
#> [2,] 0.8553641 0.9863925 0.9043802 0.4544001 0.3233717 0.4053840
#> [3,] 0.8715926 0.9043802 1.2051138 0.7003336 0.6675460 0.3668124
#> [4,] 0.5600552 0.4544001 0.7003336 0.6360838 0.7417389 0.4958054
#> [5,] 0.7643617 0.3233717 0.6675460 0.7417389 1.1827290 0.8385546
#> [6,] 0.7481332 0.4053840 0.3668124 0.4958054 0.8385546 0.8771262
#> 
#> $number_of_observations
#> [1] 13
#> 
#> $was_mean_estimated
#> [1] TRUE
#> 
#> $delta
#> [1] 3
#> 
#> $D_matrix
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]    1    0    0    0    0    0
#> [2,]    0    1    0    0    0    0
#> [3,]    0    0    1    0    0    0
#> [4,]    0    0    0    1    0    0
#> [5,]    0    0    0    0    1    0
#> [6,]    0    0    0    0    0    1
#> 
#> $optimization_algorithm_used
#> [1] "Metropolis_Hastings"
#> 
#> $did_converge
#> NULL
#> 
#> $number_of_log_posteriori_calls
#> [1] 10
#> 
#> $whole_optimization_time
#> Time difference of 0.01689696 secs
#> 
#> $log_posteriori_calls_after_best
#> [1] 9
#> 
#> $acceptance_rate
#> [1] 0
#> 

g_map2 <- find_MAP(g, max_iter = 10, show_progress_bar = FALSE, optimizer = "hill_climbing")
summary(g_map2)
#> The optimized `gips` object.
#> 
#> Permutation:
#>  (1,4)(2,5)(3,6)
#> 
#> Log_posteriori:
#>  -5.594701
#> 
#> Times more likely than starting permutation:
#> 369.906
#> 
#> Number of observations:
#>  13
#> 
#> The mean in `S` matrix was estimated.
#> Therefore, one degree of freedom was lost.
#> There is 12 degrees of freedom left.
#> 
#> n0:
#>  4
#> 
#> Number of observations is bigger than n0 for this permutaion,
#> so the gips model based on the found permutation does exist.
#> 
#> --------------------------------------------------------------------------------
#> Optimization algorithm:
#>  hill_climbing did converge
#> 
#> Number of log_posteriori calls:
#>  61
#> 
#> Optimization time:
#>  0.09211397 secs
#> 
#> Log_posteriori calls after the found permutation:
#>  27
# ================================================================================
S <- matrix(c(1, 0.5, 0.5, 2), nrow = 2, byrow = TRUE)
g <- gips(S, 10)
print(summary(g))
#> The unoptimized `gips` object.
#> 
#> Permutation:
#>  ()
#> 
#> Log_posteriori:
#>  -16.56396
#> 
#> Times more likely than identity permutation:
#> 1
#> 
#> Number of observations:
#>  10
#> 
#> The mean in `S` matrix was estimated.
#> Therefore, one degree of freedom was lost.
#> There is 9 degrees of freedom left.
#> 
#> n0:
#>  3
#> 
#> Number of observations is bigger than n0 for this permutaion,
#> so the gips model based on the found permutation does exist.