Summarizing the gips object

summary method for gips class.

Usage

# S3 method for gips
summary(object, ...)

# S3 method for summary.gips
print(x, ...)

Arguments

object

An object of class gips. Usually, a result of a find_MAP().

...

Further arguments passed to or from other methods.

x

An object of class summary.gips to be printed

Value

The function summary.gips() computes and returns a list of summary statistics of the given gips object. Those are:

For unoptimized gips object:

optimized - FALSE.

start_permutation - the permutation this gips represents.

start_permutation_log_posteriori - the log of the a posteriori value the start permutation has.

times_more_likely_than_id - how many more likely the start_permutation is over the identity permutation, (). It can be less than 1, meaning the identity permutation is more likely. Remember that this number can big and overflow to Inf or small and underflow to 0.

n0 - the minimum number of observations needed for the covariance matrix's maximum likelihood estimator (corresponding to a MAP) to exist. See \(C\sigma\) and n0 section in vignette("Theory", package = "gips") or in its pkgdown page.

S_matrix - the underlying matrix. This matrix will be used in calculations of the posteriori value in log_posteriori_of_gips().

number_of_observations - the number of observations that were observed for the S_matrix to be calculated. This value will be used in calculations of the posteriori value in log_posteriori_of_gips().

was_mean_estimated - given by the user while creating the gips object:

TRUE means the S parameter was the output of stats::cov() function;

FALSE means the S parameter was calculated with S = t(X) %*% X / number_of_observations.

delta, D_matrix - the hyperparameters of the Bayesian method. See the Hyperparameters section of gips() documentation.

AIC, BIC - output of AIC.gips() and BIC.gips() functions.

n_parameters - number of free parameters in the covariance matrix.

For optimized gips object:

optimized - TRUE.

found_permutation - the permutation this gips represents. The visited permutation with the biggest a posteriori value.

found_permutation_log_posteriori - the log of the a posteriori value the found permutation has.

start_permutation - the original permutation this gips represented before optimization. It is the first visited permutation.

start_permutation_log_posteriori - the log of the a posteriori value the start permutation has.

times_more_likely_than_start - how many more likely the found_permutation is over the start_permutation. It cannot be a number less than 1. Remember that this number can big and overflow to Inf.

n0 - the minimal number of observations needed for the existence of the maximum likelihood estimator (corresponding to a MAP) of the covariance matrix (see \(C\sigma\) and n0 section in vignette("Theory", package = "gips") or in its pkgdown page).

S_matrix - the underlying matrix. This matrix will be used in calculations of the posteriori value in log_posteriori_of_gips().

number_of_observations - the number of observations that were observed for the S_matrix to be calculated. This value will be used in calculations of the posteriori value in log_posteriori_of_gips().

was_mean_estimated - given by the user while creating the gips object:

TRUE means the S parameter was output of the stats::cov() function;

FALSE means the S parameter was calculated with S = t(X) %*% X / number_of_observations.

delta, D_matrix - the hyperparameters of the Bayesian method. See the Hyperparameters section of gips() documentation.

AIC, BIC - output of AIC.gips() and BIC.gips() functions.

n_parameters - number of free parameters in the covariance matrix.

optimization_algorithm_used - all used optimization algorithms in order (one could start optimization with "MH", and then do an "HC").

did_converge - a boolean, did the last used algorithm converge.

number_of_log_posteriori_calls - how many times was the log_posteriori_of_gips() function called during the optimization.

whole_optimization_time - how long was the optimization process; the sum of all optimization times (when there were multiple).

log_posteriori_calls_after_best - how many times was the log_posteriori_of_gips() function called after the found_permutation; in other words, how long ago could the optimization be stopped and have the same result. If this value is small, consider running find_MAP() again with optimizer = "continue". For optimizer = "BF", it is NULL.

acceptance_rate - only interesting for optimizer = "MH". How often was the algorithm accepting the change of permutation in an iteration.

The function print.summary.gips() returns an invisible NULL.

Methods (by generic)

print(summary.gips): Printing method for class summary.gips. Prints every interesting information in a form pleasant for humans.

See also

Examples

require("MASS") # for mvrnorm()

perm_size <- 6
mu <- runif(6, -10, 10) # Assume we don't know the mean
sigma_matrix <- matrix(
  data = c(
    1.0, 0.8, 0.6, 0.4, 0.6, 0.8,
    0.8, 1.0, 0.8, 0.6, 0.4, 0.6,
    0.6, 0.8, 1.0, 0.8, 0.6, 0.4,
    0.4, 0.6, 0.8, 1.0, 0.8, 0.6,
    0.6, 0.4, 0.6, 0.8, 1.0, 0.8,
    0.8, 0.6, 0.4, 0.6, 0.8, 1.0
  ),
  nrow = perm_size, byrow = TRUE
) # sigma_matrix is a matrix invariant under permutation (1,2,3,4,5,6)
number_of_observations <- 13
Z <- MASS::mvrnorm(number_of_observations, mu = mu, Sigma = sigma_matrix)
S <- cov(Z) # Assume we have to estimate the mean

g <- gips(S, number_of_observations)

g_map <- find_MAP(g, max_iter = 10, show_progress_bar = FALSE, optimizer = "Metropolis_Hastings")
unclass(summary(g_map))
#> $optimized
#> [1] TRUE
#> 
#> $found_permutation
#> [1] (25)
#> 
#> $found_permutation_log_posteriori
#> [1] -20.18477
#> 
#> $start_permutation
#> [1] ()
#> 
#> $start_permutation_log_posteriori
#> [1] -24.21009
#> 
#> $times_more_likely_than_start
#> [1] 55.9983
#> 
#> $log_times_more_likely_than_start
#> [1] 4.025321
#> 
#> $n0
#> [1] 6
#> 
#> $S_matrix
#>           [,1]      [,2]      [,3]      [,4]      [,5]      [,6]
#> [1,] 1.1391797 0.7325764 0.2125409 0.1760662 0.5826694 1.1027049
#> [2,] 0.7325764 0.9269523 0.5084455 0.3934551 0.1990792 0.6175860
#> [3,] 0.2125409 0.5084455 0.8109065 0.6123964 0.3164918 0.0140308
#> [4,] 0.1760662 0.3934551 0.6123964 0.8840775 0.6666886 0.4477473
#> [5,] 0.5826694 0.1990792 0.3164918 0.6666886 1.0502788 0.9328662
#> [6,] 1.1027049 0.6175860 0.0140308 0.4477473 0.9328662 1.5364214
#> 
#> $number_of_observations
#> [1] 13
#> 
#> $was_mean_estimated
#> [1] TRUE
#> 
#> $delta
#> [1] 3
#> 
#> $D_matrix
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]
#> [1,] 1.057969 0.000000 0.000000 0.000000 0.000000 0.000000
#> [2,] 0.000000 1.057969 0.000000 0.000000 0.000000 0.000000
#> [3,] 0.000000 0.000000 1.057969 0.000000 0.000000 0.000000
#> [4,] 0.000000 0.000000 0.000000 1.057969 0.000000 0.000000
#> [5,] 0.000000 0.000000 0.000000 0.000000 1.057969 0.000000
#> [6,] 0.000000 0.000000 0.000000 0.000000 0.000000 1.057969
#> 
#> $n_parameters
#> [1] 16
#> 
#> $AIC
#> [1] -635.937
#> 
#> $BIC
#> [1] -626.8978
#> 
#> $optimization_algorithm_used
#> [1] "Metropolis_Hastings"
#> 
#> $did_converge
#> NULL
#> 
#> $number_of_log_posteriori_calls
#> [1] 10
#> 
#> $whole_optimization_time
#> Time difference of 0.02067375 secs
#> 
#> $log_posteriori_calls_after_best
#> [1] 3
#> 
#> $acceptance_rate
#> [1] 0.1
#> 

g_map2 <- find_MAP(g, max_iter = 10, show_progress_bar = FALSE, optimizer = "hill_climbing")
summary(g_map2)
#> The optimized `gips` object.
#> 
#> Permutation:
#>  (1,4)(2,5)(3,6)
#> 
#> Log_posteriori:
#>  -15.36068
#> 
#> Times more likely than starting permutation:
#>  6970.309
#> 
#> The number of observations:
#>  13
#> 
#> The mean in the `S` matrix was estimated.
#> Therefore, one degree of freedom was lost.
#> There are 12 degrees of freedom left.
#> 
#> n0:
#>  4
#> 
#> The number of observations is bigger than n0 for this permutation,
#> so the gips model based on the found permutation does exist.
#> 
#> The number of free parameters in the covariance matrix:
#>  12
#> 
#> BIC:
#>  Inf
#> 
#> AIC:
#>  Inf
#> 
#> --------------------------------------------------------------------------------
#> Optimization algorithm:
#>  hill_climbing did converge
#> 
#> The number of log_posteriori calls:
#>  61
#> 
#> Optimization time:
#>  0.1107676 secs
#> 
#> Log_posteriori calls after the found permutation:
#>  18
# ================================================================================
S <- matrix(c(1, 0.5, 0.5, 2), nrow = 2, byrow = TRUE)
g <- gips(S, 10)
print(summary(g))
#> The unoptimized `gips` object.
#> 
#> Permutation:
#>  ()
#> 
#> Log_posteriori:
#>  -15.4837
#> 
#> The number of observations:
#>  10
#> 
#> The mean in the `S` matrix was estimated.
#> Therefore, one degree of freedom was lost.
#> There are 9 degrees of freedom left.
#> 
#> n0:
#>  3
#> 
#> The number of observations is bigger than n0 for this permutation,
#> so the gips model based on the found permutation does exist.
#> 
#> The number of free parameters in the covariance matrix:
#>  3
#> 
#> BIC:
#>  63.02608
#> 
#> AIC:
#>  62.11833